Board 백엔드

+++ Board/board/src/main/java/com/aloha/board/mapper/BaseMapper.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface BaseMapper<E> {
     List<E> list();
-    E select(int no);
+    E select(Long no);
     E selectById(String id);
     int insert(E boards);
     int update(E boards);
